requirements.txt
fandango-fuzzer>=1.0.4